Move hl-service handlers into separate files. Handlers no longer use p2putil.Read/WriteMsg.
